requirements.txt
langgraph==0.4.7
langchain==0.3.25
python-dotenv==1.1.0
langchain-openai==0.3.18
html2text==2025.4.15
langgraph-cli[inmem]